requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cssselect==1.2.0
cymem==2.0.8
dataclasses-json==0.6.6
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.14.0
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.5.0
gpt4all==2.6.0
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
html2text==2020.1.16
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.0
idna==3.7
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-ng==1.6.1
jsonpointer==2.4
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-text-splitters==0.0.2
langcodes==3.4.0
langsmith==0.1.59
language_data==1.2.0
llama-index-core==0.10.37.post1
llama-index-embeddings-langchain==0.1.2
llama-index-embeddings-openai==0.1.9
llama-index-llms-anyscale==0.1.4
llama-index-llms-langchain==0.1.3
llama-index-llms-openai==0.1.19
llama-index-readers-file==0.1.22
llama-index-readers-web==0.1.16
llamaindex-py-client==0.1.19
lxml==5.2.2
marisa-trie==1.1.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
openai==1.30.1
orjson==3.10.3
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.2
pillow==10.3.0
playwright==1.44.0
ply==3.11
preshed==3.0.9
pydantic==2.7.1
pydantic_core==2.18.2
pyee==11.1.0
Pygments==2.18.0
pypdf==4.2.0
PySocks==1.7.1
pyTelegramBotAPI==4.18.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
requests-file==2.0.0
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
selenium==4.21.0
sentence-transformers==2.7.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spider-client==0.0.11
SQLAlchemy==2.0.30
srsly==2.4.8
starlette==0.37.2
striprtf==0.0.26
sympy==1.12
telebot==0.0.5
tenacity==8.3.0
thinc==8.2.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.41.0
trio==0.25.1
trio-websocket==0.11.1
triton==2.3.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
wasabi==1.1.2
watchfiles==0.21.0
weasel==0.3.4
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
